requirements/development.txt
pytest==3.0.3
pytest-tornado==0.4.5
pytest-dotenv==0.1.0
pytest-cov==2.3.1
flake8==3.0.4
pytest-capturelog==0.7
-r production.txt
